+++ libs/newsletters-data-client/src/lib/newsletter-data-type.ts
@@ -4,7 +4,7 @@ import {
 	kebabCasedString,
 	nonEmptyString,
 	underscoreCasedString,
-} from './schema-helpers';
+} from './zod-helpers/schema-helpers';
 
 export const themeEnumSchema = z.enum([
 	'',
